Move refresh token rotation (#77)
* refactor(auth): move refresh token rotation to jwt callback

* chore(middleware): add comments regarding branches
